+++ solidity/contracts/ContractIds.sol
@@ -17,10 +17,11 @@ contract ContractIds {
     bytes32 public constant BANCOR_CONVERTER_UPGRADER = "BancorConverterUpgrader";
     bytes32 public constant BANCOR_CONVERTER_FACTORY = "BancorConverterFactory";
 
-    // Ids of BNT converter and BNT token
+    // BNT core
     bytes32 public constant BNT_TOKEN = "BNTToken";
     bytes32 public constant BNT_CONVERTER = "BNTConverter";
 
-    // Id of BancorX contract
+    // BancorX
     bytes32 public constant BANCOR_X = "BancorX";
+    bytes32 public constant BANCOR_X_UPGRADER = "BancorXUpgrader";
 }