+++ solidity/contracts/IBancorNetwork.sol
@@ -7,6 +7,20 @@ import './token/interfaces/IERC20Token.sol';
 contract IBancorNetwork {
     function convert(IERC20Token[] _path, uint256 _amount, uint256 _minReturn) public payable returns (uint256);
     function convertFor(IERC20Token[] _path, uint256 _amount, uint256 _minReturn, address _for) public payable returns (uint256);
+    
+    function convertForPrioritized3(
+        IERC20Token[] _path,
+        uint256 _amount,
+        uint256 _minReturn,
+        address _for,
+        uint256 _customVal,
+        uint256 _block,
+        uint8 _v,
+        bytes32 _r,
+        bytes32 _s
+    ) public payable returns (uint256);
+    
+    // deprecated, backward compatibility
     function convertForPrioritized2(
         IERC20Token[] _path,
         uint256 _amount,
@@ -15,8 +29,8 @@ contract IBancorNetwork {
         uint256 _block,
         uint8 _v,
         bytes32 _r,
-        bytes32 _s)
-        public payable returns (uint256);
+        bytes32 _s
+    ) public payable returns (uint256);
 
     // deprecated, backward compatibility
     function convertForPrioritized(
@@ -28,6 +42,6 @@ contract IBancorNetwork {
         uint256 _nonce,
         uint8 _v,
         bytes32 _r,
-        bytes32 _s)
-        public payable returns (uint256);
+        bytes32 _s
+    ) public payable returns (uint256);
 }