+++ solidity/contracts/converter/interfaces/IBancorConverterUpgrader.sol
@@ -6,4 +6,5 @@ import './IBancorConverter.sol';
 */
 contract IBancorConverterUpgrader {
     function upgrade(bytes32 _version) public;
+    function upgrade(uint16 _version) public;
 }