+++ solidity/migrations/2_deploy_contracts.js
@@ -13,6 +13,8 @@ const Whitelist = artifacts.require('Whitelist.sol');
 const SmartToken = artifacts.require('SmartToken.sol');
 const SmartTokenController = artifacts.require('SmartTokenController.sol');
 const BancorNetwork = artifacts.require('BancorNetwork.sol');
+const BancorX = artifacts.require('BancorX.sol');
+const XTransferRerouter = artifacts.require('XTransferRerouter.sol');
 const BancorFormula = artifacts.require('BancorFormula.sol');
 const BancorGasPriceLimit = artifacts.require('BancorGasPriceLimit.sol');
 const BancorConverter = artifacts.require('BancorConverter.sol');
@@ -42,5 +44,6 @@ module.exports = async function(deployer, network, accounts) {
         await deployer.deploy(BancorConverterUpgrader, ContractRegistry.address);
 
         deployer.deploy(CrowdsaleController, SmartToken.address, 4102444800, '0x125', '0x126', 1);
+        deployer.deploy(XTransferRerouter, true);
     }
 };